fix(terminal): unify keyboard handler — prevent prompt-nav from overwriting copy/paste
CRITICAL: xterm.js's attachCustomKeyEventHandler is a setter, not additive.
The S5 prompt-navigation useEffect in TerminalView silently replaced the
existing handler from useTerminal.ts that provides Cmd+C/V/A, font zoom,
and highlight toggle.

Fix: move Cmd+↑/↓ prompt-navigation logic into useTerminal.ts's single
attachCustomKeyEventHandler call. Remove the duplicate handler from
TerminalView.tsx.

Also: track terminal rows in React state (via onResize) so CommandGutter
uses up-to-date viewport dimensions after panel resize.

Fixes: keyboard handler overwrite (CRITICAL, Opus+GPT dual-model agreement)
Fixes: rows state for resize (MED, GPT)
